+++ src/main/java/com/frejdh/util/environment/FileUtils.java
@@ -1,9 +1,16 @@
 package com.frejdh.util.environment;
 
 import java.io.ByteArrayOutputStream;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.FileSystems;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
 
 class FileUtils {
 
@@ -17,12 +24,23 @@ class FileUtils {
 	}
 
 	/**
+	 * Load a file as an InputStream
+	 * @param absolutePath The relative path from the resource directory
+	 * @return An InputStream or null
+	 */
+	static InputStream getAbsoluteFileAsStream(String absolutePath) throws IOException {
+		return new FileInputStream(absolutePath);
+	}
+
+	/**
 	 * Load a file as a string.
-	 * @param relativePath The relative path from the resource directory
+	 * @param filePath Absolute path or relative path based on the resource directory
 	 * @return A string or null if the file couldn't be loaded
 	 */
-	static String getResourceFile(String relativePath) {
-		try (InputStream inputStream = getResourceFileAsStream(relativePath)) {
+	static String getResourceFile(String filePath) {
+		boolean isAbsolutePath = Paths.get(filePath).isAbsolute();
+
+		try (InputStream inputStream = (isAbsolutePath ? getAbsoluteFileAsStream(filePath) : getResourceFileAsStream(filePath))) {
 			ByteArrayOutputStream result = new ByteArrayOutputStream();
 			byte[] buffer = new byte[4 * 0x400]; // 4KB
 			int length;
@@ -36,4 +54,5 @@ class FileUtils {
 		}
 
 	}
+
 }